Fixed bug when writing pixel values to file

+++ src/Pixel.h
@@ -11,5 +11,6 @@ struct Pixel
 
 void operator<<(std::ostream& ostream, const Pixel& pixel)
 {
-    ostream << pixel.r << ' ' << pixel.g << ' ' << pixel.b << '\n';
+    ostream << static_cast<int>(pixel.r) << ' ' << static_cast<int>(pixel.g) << ' '
+            << static_cast<int>(pixel.b) << '\n';
 }
